feat(analysis): moved ast and protocol to analysis
