+++ tools/spxls/internal/analysis/ast/astutil/util.go
