+++ tools/spxls/internal/analysis/ast/inspector/inspector.go
