+++ tools/spxls/internal/analysis/ast/inspector/typeof.go
