+++ tools/spxls/internal/analysis/protocol/analysis.go
