+++ tools/spxls/internal/analysis/protocol/diagnosic.go
