+++ tools/spxls/internal/analysis/protocol/tsprotocol.go
